コンポーネントに切る
