+++ front/src/hooks/button/index.ts
@@ -1,0 +1,1 @@
+export { useButtonBookmark } from '@/hooks/button/useButtonBookmark'
